Reshape Data from Wide to Long Format for Cross-Lagged Analysis

This function takes a wide-format data frame containing variables named 'x1', 'x2', ..., 'xn', and 'y1', 'y2', ..., 'yn', and reshapes it into a long format suitable for cross-lagged analysis. The resulting data frame includes columns for individual ID, wave, x, y, xlag (lagged x), and ylag (lagged y).

reshape_long_sim_cr(data = dat)

Arguments

data

A wide-format data frame containing variables named 'x1', 'x2', ..., 'xn', and 'y1', 'y2', ..., 'yn'.

Value

A long-format data frame with columns 'id', 'wave', 'x', 'y', 'xlag', and 'ylag'.

Examples

dat = simulate_observed_clr(waves = 3)$data

head(dat)
#>           x1         y1         x2         y2          x3         y3          u
#> 1  1.7144916  0.8614951  1.5264989  4.4485005  3.75386514  7.9630228  1.1424622
#> 2  0.1143434  0.2500249 -0.4372785 -0.5559827 -1.14803773 -1.6076698  0.2911890
#> 3  1.5387550  4.3053445  3.3545445  8.1289103  5.27087764 13.4102381  1.4291709
#> 4 -0.4218234  0.4603176 -0.8253434  0.3380870 -0.07709191 -0.1168885  1.9885765
#> 5 -1.1323158 -1.8615668 -1.9465797 -3.1665720 -1.02628881 -2.1571460  0.3086371
#> 6 -0.6821321 -2.4582491 -0.6661364 -3.8794321 -3.48080850 -6.1323960 -0.2553419
#>             z
#> 1  0.66627536
#> 2 -0.01405512
#> 3  2.30758462
#> 4  0.17268879
#> 5 -1.52997129
#> 6 -0.78817437