Simulate data from a latent change model, two variables.

Create a simulated dataset for two variable, based on latent change score parameters. This follows McArdle's review "Latent Variable Modeling of Differences and Changes with Longitudinal data. McArdle JJ. Latent variable modeling of differences and changes with longitudinal data. Annu Rev Psychol. 2009;60:577-605. doi: 10.1146/annurev.psych.60.110707.163612. PMID: 18817479.

simulate_latent_change_multiple(
  waves = 10,
  latent.true.variance_x = 1,
  observed.residual.var_x = 0.1,
  change.mean_x = 0.01,
  constant.change.var_x = 0.1,
  proportion.change_x = 0.02,
  cov.change.true_x = 0.1,
  latent.true.variance_y = 1,
  observed.residual.var_y = 0.1,
  change.mean_y = 0.1,
  constant.change.var_y = 0.5,
  proportion.change_y = 0.3,
  cov.change.true_y = 0.1,
  coupling_xy = 0.1,
  coupling_yx = 0.1,
  cov.xy.wave1 = 0.1,
  cov.general = 0.1,
  cov.x.general = 0.1,
  cov.y.general = 0.1,
  ...
)

Arguments

waves

The number of waves (time points) in the model.

latent.true.variance_x

The true score variance, or the variance of the latent. Variable x.

observed.residual.var_x

The residual variance; the variance of the observed data, from waves 2 through K. Variable x.

change.mean_x

The mean of the latent change score factor. Variable x.

constant.change.var_x

Two types of change are modeled, proportional and constant. This is the constant change variance. Variable x.

proportion.change_x

The proportional change estimate. This is the proportion of the previous wave's true score that is added to the current wave's true score. Variable x.

cov.change.true_x

The covariance between the latent change and initial true score, the common factor at wave 1. Variable x.

latent.true.variance_y

The true score variance, or the variance of the latent. Variable y.

observed.residual.var_y

The residual variance; the variance of the observed data, from waves 2 through K. Variable y.

change.mean_y

The mean of the latent change score factor. Variable y.

constant.change.var_y

Two types of change are modeled, proportional and constant. This is the constant change variance. Variable y.

proportion.change_y

The proportional change estimate. This is the proportion of the previous wave's true score that is added to the current wave's true score. Variable y.

cov.change.true_y

The covariance between the latent change and initial true score, the common factor at wave 1. Variable y.

cov.xy.wave1

The covariance between x and y at wave 1

cov.general

The covariance between the general change factors for x and y.

slope.cov

...

Additional arguments to pass to the `lavaan::simulateData` function.

coupling.xy

The coupling score -- predicted effect of y on change in x

Value

A list containing two elements: * `model`: The Lavaan model syntax used for data simulation. * `data`: The simulated data in a data frame format.