Simulate data from a latent change model, one variable.

Create a simulated dataset for one variable, based on latent change score parameters. This follows McArdle's review "Latent Variable Modeling of Differences and Changes with Longitudinal data. McArdle JJ. Latent variable modeling of differences and changes with longitudinal data. Annu Rev Psychol. 2009;60:577-605. doi: 10.1146/annurev.psych.60.110707.163612. PMID: 18817479.

Usage

simulate_latent_change_single(
  waves = 10,
  latent.true.variance = 1,
  observed.residual.var = 0.1,
  change.mean = 0.1,
  constant.change.var = 0.5,
  proportion.change = 0.02,
  cov.change.true = 0.1,
  ...
)

Arguments

waves

The number of waves (time points) in the model.

latent.true.variance

The true score variance, or the variance of the latent

observed.residual.var

The residual variance; the variance of the observed data, from waves 2 through K

change.mean

The mean of the latent change score factor.

constant.change.var

Two types of change are modeled, proportional and constant. This is the constant change variance.

proportion.change

The proportional change estimate. This is the proportion of the previous wave's true score that is added to the current wave's true score.

cov.change.true

The covariance between the latent change and initial true score, the common factor at wave 1

...

Additional arguments to pass to the `lavaan::simulateData` function.

Value

A list containing two elements: * `model`: The Lavaan model syntax used for data simulation. * `data`: The simulated data in a data frame format.