Simulate data from an observed, random intercept cross-lagged regression (RI-CLR) model

Create a simulated dataset of a cross-lagged model with a specified number of waves and structural parameters.

Usage

simulate_riclpm(
  waves = 10,
  stability.p = 0.9,
  stability.q = 0.9,
  cross.p = 0.1,
  cross.q = 0.1,
  variance.p = 1,
  variance.q = 1,
  cov.pq = 0.5,
  beta.u = 0.5,
  ...
)

Arguments

waves

The number of waves (time points) in the model.

stability.p

The stability parameter for the x variable (autoregressive effect).

stability.q

The stability parameter for the y variable (autoregressive effect).

cross.p

The cross-lagged effect of y on x at the next time point.

cross.q

The cross-lagged effect of x on y at the next time point.

cov.pq

The covariance between x and y within the same time point.

...

Additional arguments to pass to the `lavaan::simulateData` function.

Value

A list containing two elements: * `model`: The Lavaan model syntax used for data simulation. * `data`: The simulated data in a data frame format.